requirements.txt
fastmcp
autogen-agentchat>=0.4
autogen-ext[mcp,openai]
PyGithub
python-dotenv
httpx
